Improved boundary_terms and renamed it
